bugfix: don't rely on Process.Wait(), instead use syscall.
os/exec: Do not assume process is done after Wait() returns.

Currently the Wait() implementation assumes that once the Wait4() syscall returns, the process has exited. This may not always be the case, and this assumption prevents the process from being killed by `proc.Kill()`. In situations such as a process tracing another process using the Ptrace() family of syscalls, Wait() may be called several times during the life of the traced process. It is not safe to assume that the process has exited. This assumption must be validated before calling `p.setDone()`.